requirements.txt
anthropic
duckdb
pandas
fastapi
uvicorn
python-multipart
python-dotenv
thefuzz
sentence-transformers
scikit-learn
pytest
pytest-asyncio